task-driver: utils: Refactor helpers into submodules
