+++ workers/task-driver/src/tasks/create_new_wallet.rs
@@ -32,7 +32,8 @@ use util::err_str;
 
 use crate::task_state::StateWrapper;
 use crate::traits::{Descriptor, Task, TaskContext, TaskError, TaskState};
-use crate::utils::validity_proofs::{enqueue_proof_job, find_merkle_path_with_tx};
+use crate::utils::enqueue_proof_job;
+use crate::utils::merkle_path::find_merkle_path_with_tx;
 
 /// The task name to display when logging
 const NEW_WALLET_TASK_NAME: &str = "create-new-wallet";
