+++ workers/task-driver/src/tasks/pay_relayer_fee.rs
@@ -24,9 +24,9 @@ use util::err_str;
 
 use crate::task_state::StateWrapper;
 use crate::traits::{Descriptor, Task, TaskContext, TaskError, TaskState};
-use crate::utils::validity_proofs::{
-    enqueue_proof_job, find_merkle_path_with_tx, update_wallet_validity_proofs,
-};
+use crate::utils::enqueue_proof_job;
+use crate::utils::merkle_path::find_merkle_path_with_tx;
+use crate::utils::validity_proofs::update_wallet_validity_proofs;
 
 use super::{ERR_BALANCE_MISSING, ERR_NO_MERKLE_PROOF, ERR_WALLET_MISSING};
