+++ workers/task-driver/src/tasks/settle_match_internal.rs
@@ -6,8 +6,9 @@ use std::fmt::{Display, Formatter, Result as FmtResult};
 use crate::task_state::StateWrapper;
 use crate::traits::{Descriptor, Task, TaskContext, TaskError, TaskState};
 use crate::utils::order_states::{record_order_fill, transition_order_settling};
-use crate::utils::validity_proofs::{
-    enqueue_proof_job, find_merkle_path_with_tx, update_wallet_validity_proofs,
+use crate::utils::{
+    enqueue_proof_job, merkle_path::find_merkle_path_with_tx,
+    validity_proofs::update_wallet_validity_proofs,
 };
 use alloy::rpc::types::TransactionReceipt;
 use async_trait::async_trait;
